fix(striker-ui): remove undefined hosts after host sanitize
The back-end is currently ignoring requests for status on the DR
host(s) because we don't need to show DR status in the UI (for now).
The front-end gets undefined when trying to match a DR host array index
to the AnvilStatusHost[].

This change filters out the undefined item(s) to prevent crashing.